Improvements to building validations
