+++ src/validationHandler/index.ts
@@ -42,7 +42,7 @@ export const validateIdObl = [
     .withMessage('Id must be a number')
     .bail()
     .notEmpty()
-    .withMessage('Cannot be empty')
+    .withMessage('Id cannot be empty')
     .bail(),
 ];
 export const validateNameObl = [
@@ -77,7 +77,10 @@ export const validateDescription = [
 
 export const validateDescriptionObl = [
   ...validateDescription,
-  check('description').notEmpty().withMessage('Cannot be empty').bail(),
+  check('description')
+    .notEmpty()
+    .withMessage('Description cannot be empty')
+    .bail(),
 ];
 
 export const validatePriorityMustBeNumber = [
@@ -91,33 +94,6 @@ export const validateAddEquipment = [
   check('isMovable')
     .matches(/^[01]$/)
     .withMessage('isMovable needs to be 1 = can be moved, 0 = cannot be moved.')
-    .bail(),
-];
-
-/* ---- BUILDING ---- */
-export const validateBuildingPost = [
-  ...validateNameObl,
-  ...validateDescriptionObl,
-];
-
-// This a bit different as body can have multiple objects,
-// => MultiPost!!!
-export const validateBuildingMultiPost = [
-  body('*.name')
-    .isLength({ min: 2, max: 255 })
-    .withMessage('Must be between 2-255 characters long')
-    .bail()
-    .matches(/^[A-Za-zäöåÄÖÅ0-9\s-]*$/)
-    .withMessage('Must contain only letters, numbers and -')
-    .bail()
-    .notEmpty()
-    .withMessage('Cannot be empty')
-    .bail(),
-  body('*.description')
-    .isLength({ min: 2, max: 255 })
-    .withMessage('Must be between 2-255 characters long')
-    .matches(/^[A-Za-zäöåÄÖÅ0-9\s-]*$/)
-    .withMessage('Must contain only letters, numbers and -')
     .bail(),
 ];
